some changes in exam and class entity

+++ parg-service/src/main/java/io/github/kiransr99/parg/repository/PhysicalTestPerformanceMetricRepository.java
@@ -3,5 +3,5 @@ package io.github.kiransr99.parg.repository;
 import io.github.kiransr99.parg.entity.PhysicalTestPerformanceMetric;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhysicalTestPerformanceMetricRepository extends JpaRepository<PhysicalTestPerformanceMetric, Integer> {
+public interface PhysicalTestPerformanceMetricRepository extends JpaRepository<PhysicalTestPerformanceMetric, Long> {
 }
